requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.1
Django==4.2
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
environs==11.0.0
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
marshmallow==3.22.0
multidict==6.1.0
mysql-connector-python==9.0.0
packaging==24.1
pillow==10.4.0
protobuf==3.20.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
XlsxWriter==3.2.0
yarl==1.11.1
